Add files via upload
This is a buggy but possible starting point for a program that has a run-time error that all grades are over 100. Also, a flowchart that is about 1/3 complete and doesn't exactly match the program uploaded. I also included a swimlane workflow diagram made on Lucidchart for the Project Documentation part of the project.