Add files via upload
So, the change that I made to comment out the line that subtracts one from the quotient when calculating the average, if 5 didn't go into the total score completely, solves the average being off by one in some test runs (including with the required test data). However, it now rounds up all fractional scores. For example, if the average on a calculator is 84.4, it will round up to 85. Without commenting out that line, it truncates fractional parts, but the average is one less in some test runs. I don't think this is a show-stopper, and if we can fix it before the project is due, that would be awesome. But, I don't think it's 100% necessary. I updated the comments at the top of the LC-3 file under "Side effects" to reflect this behavior of the program.